requirements.txt
streamlit>=1.34,<2
PyMuPDF>=1.24,<2
Pillow>=10,<12
numpy>=1.26,<3